refactor!: rename the transaction and lock stages to Services\Pipeline
The internal transaction and lock pipeline stages lived in Services\Concerns and were named TransactionConcern/LockConcern, which read as implementations of the user-facing ServiceConcern contract - which they are not (they carry a bespoke wrap() signature and a fixed lock-then-transaction ordering enforced by ServiceRunner, and do not implement ServiceConcern). They move to SineMacula\ApiToolkit\Services\Pipeline as TransactionStage and LockStage so Services\Concerns no longer holds them and the naming reflects their role. Method bodies, the wrap() signatures and the lock-then-transaction ordering are byte-for-byte unchanged; only the class names, namespace and file paths change. ServiceRunner and the tests are updated; the empty Services\Concerns directory is removed.

BREAKING CHANGE: the internal Services\Concerns\TransactionConcern and LockConcern stages are now Services\Pipeline\TransactionStage and LockStage.
